Added new text below the note based on suggestion from Holly
Here is the text - In the 2011 National Survey of State Vocational
Rehabilitation Agencies we asked VR agencies to report where they are
located within the government structure of their state—their Designated
State Agency (DSA) or “parent agency.” Based on survey responses to this
question we categorized VR agencies into 1 of 5 categories:
Labor/Workforce, Independent (they do not have a parent agency),
Human/Social Services, Education, or Other. We paired this data with
RSA-911 rehabilitation rate data to analyze rehabilitation rates based
on parent agency. Rehabilitation rate is the ratio of successful
closures (Status 26) relative to the sum of successful closures and
people not rehabilitated with an IPE [Status 26/(Status 26 + Status
28)]. We created two charts, one for general and combined agencies and
one for blind agencies.